+++ hlx_statics/blocks/code/code.js
@@ -1,4 +1,4 @@
-import decoratePreformattedCode, { applyLanguageDirectives, extractLanguageDirectives } from '../../components/code.js';
+import decoratePreformattedCode, { applyDataAttributesFromCodeClasses, applyLanguageDirectives, extractLanguageDirectives } from '../../components/code.js';
 
 export default function decorate(block) {
   const code = block.querySelector('code');
@@ -10,37 +10,7 @@ export default function decorate(block) {
     code.parentElement.replaceChild(pre, code);
     pre.appendChild(code);
   }
-  // Process each class from code element
-  [...code.classList].forEach(cls => {
-    if (cls.includes('-data')) {
-      const parts = cls.split('-data');
-      const languagePart = parts.find(item => item.includes('language-'));
-
-      // Add data attributes from class parts
-      parts.forEach(item => {
-        if (!item.includes('language-') && item.includes('=')) {
-          const match = item.match(/^-?([^=]+)="?([^"]*)"?$/);
-          if (match) {
-            const attrName = `data-${match[1]}`;
-            const attrValue = match[2];
-            pre.setAttribute(attrName, attrValue);
-            code.setAttribute(attrName, attrValue);
-          }
-        }
-      });
-
-      // Remove the messy class and add clean language class
-      code.classList.remove(cls);
-      pre.classList.remove(cls);
-      if (languagePart) {
-        const cleanClass = languagePart.trim();
-        code.classList.add(cleanClass);
-        pre.classList.add(cleanClass);
-      }
-    } else {
-      pre.classList.add(cls);
-    }
-  });
+  applyDataAttributesFromCodeClasses(pre, code);
 
   const languageClass = [...pre.classList].find(cls => cls.startsWith('language-'));
   let language = languageClass ? languageClass.replace('language-', '') : 'none';
